requirements.txt
streamlit>=1.37.0
mistralai>=1.1.0
python-dotenv>=1.0.1
python-docx>=1.1.0
